PHP cleanup and linting
- Moving the main class into its own file
- Adding new docblocks
- Removing `countrySelectorEnabled` from the `form_fields` endpoint
